Update docs to include new changes (#47)
This is squashed merged with several commits:

* Adds AUTHORS.md file.

We received a patch from Hari Nair (CommScope), since he is not visible in the Github contributors lists, we would like to give him credit with authors document.

* Updates the documentation with latest changes.

Reworked the documentation to incorporate the new Kafka import feature.
Similarly, added several new descriptions on user provided properties.

* Adds developer guide documentation.

This shows general development workflow of the project. It provides instructions on how 
to perform build, test and release actions.

* Adds coveralls code coverage plugin.

So far we use the codecov plugin, but it is usually flakey. This coveralls addition should show 
additional code coverage results.