Correctly import Parquet timestamps encoded as milliseconds since epoch.
So far, we could import Parquet timestamp values that are encoded as
`INT96` binary type. This type encodes timestamps as 12-byte binary with
nanosecond precision.

However, it is possible to also encode timestamp as milliseconds since
epoch using `INT64` type.

For example,

```
message test_timestamps {
  required int64 col_long;
  required int64 col_timestamp (TIMESTAMP_MILLIS);
}
```

the first column contains Long value and the second column millis since
epoch as Long value, but should be imported as Timestamp value.

Previously the second column value assumed to be INT96 binary which was
a bug.

Fixes #52.