Adds integration tests for Kinesis.
Porting the project to Java 11 due to an introduction of
Exasol docker test-containers framework for integration testing.

Integration tests are based on the communication between Exasol
test-containers and Localstack with Kinesis Service.

Adds Mockito Extension because we were getting NullPointerExceptions
when mocking the final methods of a class. We learned that by default
mockito does not support mocking final methods. In order to enable it,
mockito extension should be added to the test resources.

Apply suggestions from code review.

References:

* https://github.com/exasol/exasol-testcontainers
* https://www.testcontainers.org/modules/localstack/
* https://github.com/mockito/mockito/wiki/What%27s-new-in-Mockito-2#mock-the-unmockable-opt-in-mocking-of-final-classesmethods

Co-Authored-By: Sebastian Bär <sebastian.baer@exasol.com>