#117: Added exasol-testcontainers for integration tests (#119)
## Added data importer integration tests

## Fixed timestamp integration tests

The Parquet and Avro format store long value (the number of milliseconds
since epoch) as timestamp value. Thus no timezone information is encoded
when writing.

However, when we import the files using Exasol UDFs, we construct a Java
`java.sql.Timestamp` object from the milliseconds since epoch, and emit
them into table. At this moment, the Java Timestamp uses the JVM
timezone. It is usually `Europe/Berlin` for instance for the Exasol
docker container. You can the Exasol system timezone using `DBTIMEZONE`
(`SELECT DBTIMEZONE`).

This introduces a difference integration tests, timestamps written in
UTC is returned in Europe/Berlin timezone. Therefore, we shift the
expected values to database timezone.

This issue does not occur in the Orc format because it also encodes the
timezone information in the file so that Orc readers uses that timezone
(https://github.com/apache/orc/pull/249).

## Added nested suites with single docker container stack